Revert "ADM 373:[docs] [frontend] [backend] Generate and Export pipeline data" (#463)

+++ backend/src/main/java/heartbeat/service/report/LeadTimeForChangesCalculator.java
@@ -33,14 +33,13 @@ public class LeadTimeForChangesCalculator {
 				return new HashMap<String, Double>();
 			}
 			int times = item.getLeadTimes().size();
-
-			HashMap<Long, Long> totalDelayTime = item.getLeadTimes()
+			HashMap<Double, Double> totalDelayTime = item.getLeadTimes()
 				.stream()
 				.map(this::getDelayTimeMapWithLeadTime)
 				.reduce(new HashMap<>(), (pre, now) -> now);
 
-			double totalPrDelayTime = totalDelayTime.keySet().stream().reduce(0L, Long::sum);
-			double totalPipelineDelayTime = totalDelayTime.values().stream().reduce(0L, Long::sum);
+			double totalPrDelayTime = totalDelayTime.keySet().stream().reduce(0d, Double::sum);
+			double totalPipelineDelayTime = totalDelayTime.values().stream().reduce(0d, Double::sum);
 
 			double avgPrDelayTime = TimeUtil.convertMillisecondToMinutes(totalPrDelayTime / times);
 			double avgPipelineDelayTime = TimeUtil.convertMillisecondToMinutes(totalPipelineDelayTime / times);
@@ -76,8 +75,8 @@ public class LeadTimeForChangesCalculator {
 		return new LeadTimeForChanges(leadTimeForChangesOfPipelines, avgLeadTimeForChanges);
 	}
 
-	private HashMap<Long, Long> getDelayTimeMapWithLeadTime(LeadTime leadTime) {
-		HashMap<Long, Long> delayTimeMap = new HashMap<>();
+	private HashMap<Double, Double> getDelayTimeMapWithLeadTime(LeadTime leadTime) {
+		HashMap<Double, Double> delayTimeMap = new HashMap<>();
 		delayTimeMap.put(leadTime.getPrDelayTime(), leadTime.getPipelineDelayTime());
 		return delayTimeMap;
 	}
